+++ common/src/main/java/io/github/mortuusars/chalk/world/item/ChalkItem.java
@@ -4,9 +4,11 @@ import io.github.mortuusars.chalk.Chalk;
 import io.github.mortuusars.chalk.Config;
 import io.github.mortuusars.chalk.world.chalk.Mark;
 import io.github.mortuusars.chalk.world.chalk.MarkDrawingContext;
+import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.component.DataComponents;
+import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.util.FastColor;
 import net.minecraft.world.InteractionHand;
@@ -16,9 +18,12 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.component.DyedItemColor;
 import net.minecraft.world.item.context.UseOnContext;
 import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
 
 public class ChalkItem extends Item implements MarkDrawable {
     public ChalkItem(Properties properties) {
@@ -51,6 +56,13 @@ public class ChalkItem extends Item implements MarkDrawable {
         }
 
         return super.getDescriptionId(stack);
+    }
+
+    @Override
+    public void appendHoverText(ItemStack stack, TooltipContext context, List<Component> tooltipComponents, TooltipFlag tooltipFlag) {
+        if (!stack.has(DataComponents.DYED_COLOR)) {
+            tooltipComponents.add(Component.translatable("item.chalk.chalk.tooltip.dyeable").withStyle(ChatFormatting.GRAY));
+        }
     }
 
     // --